Fixed 'raw type' warnings, refactoring
