+++ SlotMachienNXT/src/io/UsbIO.java
@@ -8,7 +8,6 @@ import java.io.IOException;
 import java.io.InputStreamReader;
 import java.io.OutputStreamWriter;
 
-import slotmachien.actions.Action;
 import lejos.nxt.comm.NXTConnection;
 import lejos.nxt.comm.USB;
