+++ SlotMachienNXT/src/slotmachien/signals/ButtonSignal.java
@@ -1,5 +1,6 @@
-package observable;
+package slotmachien.signals;
 
+import observable.Signal;
 import lejos.nxt.Button;
 
 public class ButtonSignal extends Signal {
