+++ SlotMachienNXT/src/slotmachien/signals/MovedToSignal.java
@@ -1,5 +1,6 @@
-package observable;
+package slotmachien.signals;
 
+import observable.Signal;
 import slotmachien.Command;
 import slotmachien.Position;
